+++ test/node/utils/arrow-utils-test.js
@@ -22,13 +22,13 @@ import test from 'tape';
 
 import {ArrowLoader} from '@loaders.gl/arrow';
 import {ARROW_GEO_METADATA_KEY} from '@kepler.gl/constants';
-import {parseGeometryFromArrow} from '@kepler.gl/layers';
+import {parseGeometryFromArrow, GEOARROW_ENCODINGS} from '@kepler.gl/utils';
 
-// an arrow file converted from the geojson file below in arraybuffer format
+// an arrow file converted from the geojson below in arraybuffer format
 const testPointArrowRawBytes =
   '65,82,82,79,87,49,0,0,255,255,255,255,216,3,0,0,16,0,0,0,0,0,10,0,14,0,6,0,5,0,8,0,10,0,0,0,0,1,4,0,16,0,0,0,0,0,10,0,12,0,0,0,4,0,8,0,10,0,0,0,80,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,100,253,255,255,52,2,0,0,4,0,0,0,39,2,0,0,123,34,115,99,104,101,109,97,95,118,101,114,115,105,111,110,34,58,34,48,46,49,46,48,34,44,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,112,111,105,110,116,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,103,100,97,108,58,103,101,111,109,101,116,114,121,95,116,121,112,101,34,58,34,80,111,105,110,116,34,125,125,125,0,3,0,0,0,103,101,111,0,3,0,0,0,36,1,0,0,228,0,0,0,24,0,0,0,0,0,18,0,24,0,8,0,6,0,7,0,12,0,0,0,16,0,20,0,18,0,0,0,0,0,1,16,24,0,0,0,116,0,0,0,8,0,0,0,28,0,0,0,1,0,0,0,124,0,0,0,8,0,0,0,103,101,111,109,101,116,114,121,0,0,0,0,1,0,0,0,12,0,0,0,8,0,12,0,4,0,8,0,8,0,0,0,28,0,0,0,4,0,0,0,14,0,0,0,103,101,111,97,114,114,111,119,46,112,111,105,110,116,0,0,20,0,0,0,65,82,82,79,87,58,101,120,116,101,110,115,105,111,110,58,110,97,109,101,0,0,6,0,8,0,4,0,6,0,0,0,2,0,0,0,16,0,20,0,8,0,0,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,0,3,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,120,121,0,0,0,0,6,0,8,0,6,0,6,0,0,0,0,0,2,0,212,255,255,255,0,0,1,5,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,4,0,0,0,110,97,109,101,0,0,0,0,4,0,4,0,4,0,0,0,16,0,20,0,8,0,6,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,1,2,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,105,100,0,0,8,0,12,0,8,0,7,0,8,0,0,0,0,0,0,1,32,0,0,0,255,255,255,255,24,1,0,0,20,0,0,0,0,0,0,0,12,0,22,0,6,0,5,0,8,0,12,0,12,0,0,0,0,3,4,0,24,0,0,0,40,0,0,0,0,0,0,0,0,0,10,0,24,0,12,0,4,0,8,0,10,0,0,0,156,0,0,0,16,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,16,0,0,0,0,0,0,0,5,0,0,0,0,0,0,0,24,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,24,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,24,0,0,0,0,0,0,0,16,0,0,0,0,0,0,0,0,0,0,0,4,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,5,0,0,0,110,97,109,101,50,0,0,0,0,0,0,0,0,0,240,63,0,0,0,0,0,0,240,63,255,255,255,255,0,0,0,0,24,0,0,0,0,0,0,0,0,0,14,0,24,0,6,0,8,0,12,0,16,0,20,0,14,0,0,0,0,0,4,0,120,2,0,0,96,2,0,0,60,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,240,250,255,255,24,2,0,0,4,0,0,0,10,2,0,0,123,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,112,111,105,110,116,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,98,98,111,120,34,58,91,49,46,48,44,49,46,48,44,49,46,48,44,49,46,48,93,125,125,125,0,0,8,0,0,0,103,100,97,108,58,103,101,111,0,0,0,0,1,0,0,0,232,3,0,0,0,0,0,0,32,1,0,0,0,0,0,0,40,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,10,0,12,0,0,0,4,0,8,0,10,0,0,0,80,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,100,253,255,255,52,2,0,0,4,0,0,0,39,2,0,0,123,34,115,99,104,101,109,97,95,118,101,114,115,105,111,110,34,58,34,48,46,49,46,48,34,44,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,112,111,105,110,116,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,103,100,97,108,58,103,101,111,109,101,116,114,121,95,116,121,112,101,34,58,34,80,111,105,110,116,34,125,125,125,0,3,0,0,0,103,101,111,0,3,0,0,0,36,1,0,0,228,0,0,0,24,0,0,0,0,0,18,0,24,0,8,0,6,0,7,0,12,0,0,0,16,0,20,0,18,0,0,0,0,0,1,16,24,0,0,0,116,0,0,0,8,0,0,0,28,0,0,0,1,0,0,0,124,0,0,0,8,0,0,0,103,101,111,109,101,116,114,121,0,0,0,0,1,0,0,0,12,0,0,0,8,0,12,0,4,0,8,0,8,0,0,0,28,0,0,0,4,0,0,0,14,0,0,0,103,101,111,97,114,114,111,119,46,112,111,105,110,116,0,0,20,0,0,0,65,82,82,79,87,58,101,120,116,101,110,115,105,111,110,58,110,97,109,101,0,0,6,0,8,0,4,0,6,0,0,0,2,0,0,0,16,0,20,0,8,0,0,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,0,3,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,120,121,0,0,0,0,6,0,8,0,6,0,6,0,0,0,0,0,2,0,212,255,255,255,0,0,1,5,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,4,0,0,0,110,97,109,101,0,0,0,0,4,0,4,0,4,0,0,0,16,0,20,0,8,0,6,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,1,2,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,105,100,0,0,8,0,12,0,8,0,7,0,8,0,0,0,0,0,0,1,32,0,0,0,72,6,0,0,65,82,82,79,87,49';
 
-// a simple geojson file contains one point
+// a simple geojson contains one point
 const expectedPointGeojson = {
   type: 'FeatureCollection',
   features: [
@@ -46,7 +46,7 @@ const expectedPointGeojson = {
   ]
 };
 
-// a simple geojson file contains one linestring
+// a simple geojson contains one linestring
 const testLineStringArrowRawBytes =
   '65,82,82,79,87,49,0,0,255,255,255,255,24,4,0,0,16,0,0,0,0,0,10,0,14,0,6,0,5,0,8,0,10,0,0,0,0,1,4,0,16,0,0,0,0,0,10,0,12,0,0,0,4,0,8,0,10,0,0,0,92,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,88,253,255,255,64,2,0,0,4,0,0,0,49,2,0,0,123,34,115,99,104,101,109,97,95,118,101,114,115,105,111,110,34,58,34,48,46,49,46,48,34,44,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,108,105,110,101,115,116,114,105,110,103,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,103,100,97,108,58,103,101,111,109,101,116,114,121,95,116,121,112,101,34,58,34,76,105,110,101,83,116,114,105,110,103,34,125,125,125,0,0,0,3,0,0,0,103,101,111,0,3,0,0,0,84,1,0,0,20,1,0,0,24,0,0,0,0,0,18,0,24,0,8,0,6,0,7,0,12,0,0,0,16,0,20,0,18,0,0,0,0,0,1,12,24,0,0,0,116,0,0,0,8,0,0,0,28,0,0,0,1,0,0,0,104,0,0,0,8,0,0,0,103,101,111,109,101,116,114,121,0,0,0,0,1,0,0,0,12,0,0,0,8,0,12,0,4,0,8,0,8,0,0,0,32,0,0,0,4,0,0,0,19,0,0,0,103,101,111,97,114,114,111,119,46,108,105,110,101,115,116,114,105,110,103,0,20,0,0,0,65,82,82,79,87,58,101,120,116,101,110,115,105,111,110,58,110,97,109,101,0,0,0,0,100,255,255,255,96,255,255,255,0,0,1,16,20,0,0,0,32,0,0,0,4,0,0,0,1,0,0,0,44,0,0,0,4,0,0,0,105,116,101,109,0,0,6,0,8,0,4,0,6,0,0,0,2,0,0,0,16,0,20,0,8,0,0,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,0,3,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,120,121,0,0,0,0,6,0,8,0,6,0,6,0,0,0,0,0,2,0,212,255,255,255,0,0,1,5,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,4,0,0,0,110,97,109,101,0,0,0,0,4,0,4,0,4,0,0,0,16,0,20,0,8,0,6,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,1,2,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,105,100,0,0,8,0,12,0,8,0,7,0,8,0,0,0,0,0,0,1,32,0,0,0,0,0,0,0,255,255,255,255,72,1,0,0,20,0,0,0,0,0,0,0,12,0,22,0,6,0,5,0,8,0,12,0,12,0,0,0,0,3,4,0,24,0,0,0,64,0,0,0,0,0,0,0,0,0,10,0,24,0,12,0,4,0,8,0,10,0,0,0,188,0,0,0,16,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,10,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,16,0,0,0,0,0,0,0,5,0,0,0,0,0,0,0,24,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,24,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,0,0,0,0,0,0,0,32,0,0,0,0,0,0,0,0,0,0,0,5,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,5,0,0,0,110,97,109,101,49,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,240,63,0,0,0,0,0,0,240,63,255,255,255,255,0,0,0,0,20,0,0,0,0,0,14,0,24,0,6,0,8,0,12,0,16,0,20,0,14,0,0,0,0,0,4,0,120,2,0,0,100,2,0,0,64,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,228,250,255,255,28,2,0,0,4,0,0,0,15,2,0,0,123,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,108,105,110,101,115,116,114,105,110,103,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,98,98,111,120,34,58,91,48,46,48,44,48,46,48,44,49,46,48,44,49,46,48,93,125,125,125,0,8,0,0,0,103,100,97,108,58,103,101,111,0,0,0,0,1,0,0,0,40,4,0,0,0,0,0,0,80,1,0,0,0,0,0,0,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,10,0,12,0,0,0,4,0,8,0,10,0,0,0,92,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,88,253,255,255,64,2,0,0,4,0,0,0,49,2,0,0,123,34,115,99,104,101,109,97,95,118,101,114,115,105,111,110,34,58,34,48,46,49,46,48,34,44,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,108,105,110,101,115,116,114,105,110,103,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,103,100,97,108,58,103,101,111,109,101,116,114,121,95,116,121,112,101,34,58,34,76,105,110,101,83,116,114,105,110,103,34,125,125,125,0,0,0,3,0,0,0,103,101,111,0,3,0,0,0,84,1,0,0,20,1,0,0,24,0,0,0,0,0,18,0,24,0,8,0,6,0,7,0,12,0,0,0,16,0,20,0,18,0,0,0,0,0,1,12,24,0,0,0,116,0,0,0,8,0,0,0,28,0,0,0,1,0,0,0,104,0,0,0,8,0,0,0,103,101,111,109,101,116,114,121,0,0,0,0,1,0,0,0,12,0,0,0,8,0,12,0,4,0,8,0,8,0,0,0,32,0,0,0,4,0,0,0,19,0,0,0,103,101,111,97,114,114,111,119,46,108,105,110,101,115,116,114,105,110,103,0,20,0,0,0,65,82,82,79,87,58,101,120,116,101,110,115,105,111,110,58,110,97,109,101,0,0,0,0,100,255,255,255,96,255,255,255,0,0,1,16,20,0,0,0,32,0,0,0,4,0,0,0,1,0,0,0,44,0,0,0,4,0,0,0,105,116,101,109,0,0,6,0,8,0,4,0,6,0,0,0,2,0,0,0,16,0,20,0,8,0,0,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,0,3,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,120,121,0,0,0,0,6,0,8,0,6,0,6,0,0,0,0,0,2,0,212,255,255,255,0,0,1,5,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,4,0,0,0,110,97,109,101,0,0,0,0,4,0,4,0,4,0,0,0,16,0,20,0,8,0,6,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,1,2,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,105,100,0,0,8,0,12,0,8,0,7,0,8,0,0,0,0,0,0,1,32,0,0,0,128,6,0,0,65,82,82,79,87,49';
 const expectedLineStringGeoJson = {
@@ -69,7 +69,7 @@ const expectedLineStringGeoJson = {
   ]
 };
 
-// a simple geojson file contains one polygon
+// a simple geojson contains one polygon
 const testPolygonArrowRawBytes =
   '65,82,82,79,87,49,0,0,255,255,255,255,56,4,0,0,16,0,0,0,0,0,10,0,14,0,6,0,5,0,8,0,10,0,0,0,0,1,4,0,16,0,0,0,0,0,10,0,12,0,0,0,4,0,8,0,10,0,0,0,84,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,96,253,255,255,56,2,0,0,4,0,0,0,43,2,0,0,123,34,115,99,104,101,109,97,95,118,101,114,115,105,111,110,34,58,34,48,46,49,46,48,34,44,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,112,111,108,121,103,111,110,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,103,100,97,108,58,103,101,111,109,101,116,114,121,95,116,121,112,101,34,58,34,80,111,108,121,103,111,110,34,125,125,125,0,3,0,0,0,103,101,111,0,3,0,0,0,128,1,0,0,64,1,0,0,24,0,0,0,0,0,18,0,24,0,8,0,6,0,7,0,12,0,0,0,16,0,20,0,18,0,0,0,0,0,1,12,24,0,0,0,116,0,0,0,8,0,0,0,28,0,0,0,1,0,0,0,104,0,0,0,8,0,0,0,103,101,111,109,101,116,114,121,0,0,0,0,1,0,0,0,12,0,0,0,8,0,12,0,4,0,8,0,8,0,0,0,32,0,0,0,4,0,0,0,16,0,0,0,103,101,111,97,114,114,111,119,46,112,111,108,121,103,111,110,0,0,0,0,20,0,0,0,65,82,82,79,87,58,101,120,116,101,110,115,105,111,110,58,110,97,109,101,0,0,0,0,56,255,255,255,52,255,255,255,0,0,1,12,20,0,0,0,28,0,0,0,4,0,0,0,1,0,0,0,20,0,0,0,4,0,0,0,105,116,101,109,0,0,0,0,100,255,255,255,96,255,255,255,0,0,1,16,20,0,0,0,32,0,0,0,4,0,0,0,1,0,0,0,44,0,0,0,4,0,0,0,105,116,101,109,0,0,6,0,8,0,4,0,6,0,0,0,2,0,0,0,16,0,20,0,8,0,0,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,0,3,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,120,121,0,0,0,0,6,0,8,0,6,0,6,0,0,0,0,0,2,0,212,255,255,255,0,0,1,5,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,4,0,0,0,110,97,109,101,0,0,0,0,4,0,4,0,4,0,0,0,16,0,20,0,8,0,6,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,1,2,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,105,100,0,0,8,0,12,0,8,0,7,0,8,0,0,0,0,0,0,1,32,0,0,0,255,255,255,255,120,1,0,0,20,0,0,0,0,0,0,0,12,0,22,0,6,0,5,0,8,0,12,0,12,0,0,0,0,3,4,0,24,0,0,0,104,0,0,0,0,0,0,0,0,0,10,0,24,0,12,0,4,0,8,0,10,0,0,0,220,0,0,0,16,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,12,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,16,0,0,0,0,0,0,0,5,0,0,0,0,0,0,0,24,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,24,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,40,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,40,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,40,0,0,0,0,0,0,0,64,0,0,0,0,0,0,0,0,0,0,0,6,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,5,0,0,0,110,97,109,101,49,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,240,63,0,0,0,0,0,0,240,63,0,0,0,0,0,0,0,64,0,0,0,0,0,0,0,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,255,255,255,255,0,0,0,0,20,0,0,0,0,0,14,0,24,0,6,0,8,0,12,0,16,0,20,0,14,0,0,0,0,0,4,0,124,2,0,0,100,2,0,0,64,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,232,250,255,255,28,2,0,0,4,0,0,0,12,2,0,0,123,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,112,111,108,121,103,111,110,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,98,98,111,120,34,58,91,48,46,48,44,48,46,48,44,50,46,48,44,50,46,48,93,125,125,125,0,0,0,0,8,0,0,0,103,100,97,108,58,103,101,111,0,0,0,0,1,0,0,0,72,4,0,0,0,0,0,0,128,1,0,0,0,0,0,0,104,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,10,0,12,0,0,0,4,0,8,0,10,0,0,0,84,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,96,253,255,255,56,2,0,0,4,0,0,0,43,2,0,0,123,34,115,99,104,101,109,97,95,118,101,114,115,105,111,110,34,58,34,48,46,49,46,48,34,44,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,112,111,108,121,103,111,110,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,103,100,97,108,58,103,101,111,109,101,116,114,121,95,116,121,112,101,34,58,34,80,111,108,121,103,111,110,34,125,125,125,0,3,0,0,0,103,101,111,0,3,0,0,0,128,1,0,0,64,1,0,0,24,0,0,0,0,0,18,0,24,0,8,0,6,0,7,0,12,0,0,0,16,0,20,0,18,0,0,0,0,0,1,12,24,0,0,0,116,0,0,0,8,0,0,0,28,0,0,0,1,0,0,0,104,0,0,0,8,0,0,0,103,101,111,109,101,116,114,121,0,0,0,0,1,0,0,0,12,0,0,0,8,0,12,0,4,0,8,0,8,0,0,0,32,0,0,0,4,0,0,0,16,0,0,0,103,101,111,97,114,114,111,119,46,112,111,108,121,103,111,110,0,0,0,0,20,0,0,0,65,82,82,79,87,58,101,120,116,101,110,115,105,111,110,58,110,97,109,101,0,0,0,0,56,255,255,255,52,255,255,255,0,0,1,12,20,0,0,0,28,0,0,0,4,0,0,0,1,0,0,0,20,0,0,0,4,0,0,0,105,116,101,109,0,0,0,0,100,255,255,255,96,255,255,255,0,0,1,16,20,0,0,0,32,0,0,0,4,0,0,0,1,0,0,0,44,0,0,0,4,0,0,0,105,116,101,109,0,0,6,0,8,0,4,0,6,0,0,0,2,0,0,0,16,0,20,0,8,0,0,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,0,3,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,120,121,0,0,0,0,6,0,8,0,6,0,6,0,0,0,0,0,2,0,212,255,255,255,0,0,1,5,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,4,0,0,0,110,97,109,101,0,0,0,0,4,0,4,0,4,0,0,0,16,0,20,0,8,0,6,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,1,2,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,105,100,0,0,8,0,12,0,8,0,7,0,8,0,0,0,0,0,0,1,32,0,0,0,168,6,0,0,65,82,82,79,87,49';
 
@@ -97,7 +97,7 @@ const expectedPolygonGeojson = {
   ]
 };
 
-// a simple geojson file contains one MultiPoint
+// a simple geojson contains one MultiPoint
 const testMultiPointArrowRawBytes =
   '65,82,82,79,87,49,0,0,255,255,255,255,24,4,0,0,16,0,0,0,0,0,10,0,14,0,6,0,5,0,8,0,10,0,0,0,0,1,4,0,16,0,0,0,0,0,10,0,12,0,0,0,4,0,8,0,10,0,0,0,92,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,88,253,255,255,64,2,0,0,4,0,0,0,49,2,0,0,123,34,115,99,104,101,109,97,95,118,101,114,115,105,111,110,34,58,34,48,46,49,46,48,34,44,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,109,117,108,116,105,112,111,105,110,116,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,103,100,97,108,58,103,101,111,109,101,116,114,121,95,116,121,112,101,34,58,34,77,117,108,116,105,80,111,105,110,116,34,125,125,125,0,0,0,3,0,0,0,103,101,111,0,3,0,0,0,84,1,0,0,20,1,0,0,24,0,0,0,0,0,18,0,24,0,8,0,6,0,7,0,12,0,0,0,16,0,20,0,18,0,0,0,0,0,1,12,24,0,0,0,116,0,0,0,8,0,0,0,28,0,0,0,1,0,0,0,104,0,0,0,8,0,0,0,103,101,111,109,101,116,114,121,0,0,0,0,1,0,0,0,12,0,0,0,8,0,12,0,4,0,8,0,8,0,0,0,32,0,0,0,4,0,0,0,19,0,0,0,103,101,111,97,114,114,111,119,46,109,117,108,116,105,112,111,105,110,116,0,20,0,0,0,65,82,82,79,87,58,101,120,116,101,110,115,105,111,110,58,110,97,109,101,0,0,0,0,100,255,255,255,96,255,255,255,0,0,1,16,20,0,0,0,32,0,0,0,4,0,0,0,1,0,0,0,44,0,0,0,4,0,0,0,105,116,101,109,0,0,6,0,8,0,4,0,6,0,0,0,2,0,0,0,16,0,20,0,8,0,0,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,0,3,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,120,121,0,0,0,0,6,0,8,0,6,0,6,0,0,0,0,0,2,0,212,255,255,255,0,0,1,5,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,4,0,0,0,110,97,109,101,0,0,0,0,4,0,4,0,4,0,0,0,16,0,20,0,8,0,6,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,1,2,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,105,100,0,0,8,0,12,0,8,0,7,0,8,0,0,0,0,0,0,1,32,0,0,0,0,0,0,0,255,255,255,255,72,1,0,0,20,0,0,0,0,0,0,0,12,0,22,0,6,0,5,0,8,0,12,0,12,0,0,0,0,3,4,0,24,0,0,0,64,0,0,0,0,0,0,0,0,0,10,0,24,0,12,0,4,0,8,0,10,0,0,0,188,0,0,0,16,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,10,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,16,0,0,0,0,0,0,0,5,0,0,0,0,0,0,0,24,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,24,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,0,0,0,0,0,0,0,32,0,0,0,0,0,0,0,0,0,0,0,5,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,5,0,0,0,110,97,109,101,50,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,240,63,0,0,0,0,0,0,240,63,0,0,0,0,0,0,0,64,0,0,0,0,0,0,0,64,255,255,255,255,0,0,0,0,20,0,0,0,0,0,14,0,24,0,6,0,8,0,12,0,16,0,20,0,14,0,0,0,0,0,4,0,120,2,0,0,100,2,0,0,64,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,228,250,255,255,28,2,0,0,4,0,0,0,15,2,0,0,123,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,109,117,108,116,105,112,111,105,110,116,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,98,98,111,120,34,58,91,49,46,48,44,49,46,48,44,50,46,48,44,50,46,48,93,125,125,125,0,8,0,0,0,103,100,97,108,58,103,101,111,0,0,0,0,1,0,0,0,40,4,0,0,0,0,0,0,80,1,0,0,0,0,0,0,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,10,0,12,0,0,0,4,0,8,0,10,0,0,0,92,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,88,253,255,255,64,2,0,0,4,0,0,0,49,2,0,0,123,34,115,99,104,101,109,97,95,118,101,114,115,105,111,110,34,58,34,48,46,49,46,48,34,44,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,109,117,108,116,105,112,111,105,110,116,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,103,100,97,108,58,103,101,111,109,101,116,114,121,95,116,121,112,101,34,58,34,77,117,108,116,105,80,111,105,110,116,34,125,125,125,0,0,0,3,0,0,0,103,101,111,0,3,0,0,0,84,1,0,0,20,1,0,0,24,0,0,0,0,0,18,0,24,0,8,0,6,0,7,0,12,0,0,0,16,0,20,0,18,0,0,0,0,0,1,12,24,0,0,0,116,0,0,0,8,0,0,0,28,0,0,0,1,0,0,0,104,0,0,0,8,0,0,0,103,101,111,109,101,116,114,121,0,0,0,0,1,0,0,0,12,0,0,0,8,0,12,0,4,0,8,0,8,0,0,0,32,0,0,0,4,0,0,0,19,0,0,0,103,101,111,97,114,114,111,119,46,109,117,108,116,105,112,111,105,110,116,0,20,0,0,0,65,82,82,79,87,58,101,120,116,101,110,115,105,111,110,58,110,97,109,101,0,0,0,0,100,255,255,255,96,255,255,255,0,0,1,16,20,0,0,0,32,0,0,0,4,0,0,0,1,0,0,0,44,0,0,0,4,0,0,0,105,116,101,109,0,0,6,0,8,0,4,0,6,0,0,0,2,0,0,0,16,0,20,0,8,0,0,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,0,3,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,120,121,0,0,0,0,6,0,8,0,6,0,6,0,0,0,0,0,2,0,212,255,255,255,0,0,1,5,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,4,0,0,0,110,97,109,101,0,0,0,0,4,0,4,0,4,0,0,0,16,0,20,0,8,0,6,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,1,2,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,105,100,0,0,8,0,12,0,8,0,7,0,8,0,0,0,0,0,0,1,32,0,0,0,128,6,0,0,65,82,82,79,87,49';
 const expectedMultiPointGeoJson = {
@@ -120,7 +120,7 @@ const expectedMultiPointGeoJson = {
   ]
 };
 
-// a simple geojson file contains one MultiLinestring
+// a simple geojson contains one MultiLinestring
 const testMultiLineStringArrowRawBytes =
   '65,82,82,79,87,49,0,0,255,255,255,255,80,4,0,0,16,0,0,0,0,0,10,0,14,0,6,0,5,0,8,0,10,0,0,0,0,1,4,0,16,0,0,0,0,0,10,0,12,0,0,0,4,0,8,0,10,0,0,0,100,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,80,253,255,255,72,2,0,0,4,0,0,0,59,2,0,0,123,34,115,99,104,101,109,97,95,118,101,114,115,105,111,110,34,58,34,48,46,49,46,48,34,44,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,109,117,108,116,105,108,105,110,101,115,116,114,105,110,103,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,103,100,97,108,58,103,101,111,109,101,116,114,121,95,116,121,112,101,34,58,34,77,117,108,116,105,76,105,110,101,83,116,114,105,110,103,34,125,125,125,0,3,0,0,0,103,101,111,0,3,0,0,0,136,1,0,0,72,1,0,0,24,0,0,0,0,0,18,0,24,0,8,0,6,0,7,0,12,0,0,0,16,0,20,0,18,0,0,0,0,0,1,12,24,0,0,0,124,0,0,0,8,0,0,0,28,0,0,0,1,0,0,0,112,0,0,0,8,0,0,0,103,101,111,109,101,116,114,121,0,0,0,0,1,0,0,0,12,0,0,0,8,0,12,0,4,0,8,0,8,0,0,0,40,0,0,0,4,0,0,0,24,0,0,0,103,101,111,97,114,114,111,119,46,109,117,108,116,105,108,105,110,101,115,116,114,105,110,103,0,0,0,0,20,0,0,0,65,82,82,79,87,58,101,120,116,101,110,115,105,111,110,58,110,97,109,101,0,0,0,0,56,255,255,255,52,255,255,255,0,0,1,12,20,0,0,0,28,0,0,0,4,0,0,0,1,0,0,0,20,0,0,0,4,0,0,0,105,116,101,109,0,0,0,0,100,255,255,255,96,255,255,255,0,0,1,16,20,0,0,0,32,0,0,0,4,0,0,0,1,0,0,0,44,0,0,0,4,0,0,0,105,116,101,109,0,0,6,0,8,0,4,0,6,0,0,0,2,0,0,0,16,0,20,0,8,0,0,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,0,3,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,120,121,0,0,0,0,6,0,8,0,6,0,6,0,0,0,0,0,2,0,212,255,255,255,0,0,1,5,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,4,0,0,0,110,97,109,101,0,0,0,0,4,0,4,0,4,0,0,0,16,0,20,0,8,0,6,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,1,2,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,105,100,0,0,8,0,12,0,8,0,7,0,8,0,0,0,0,0,0,1,32,0,0,0,255,255,255,255,120,1,0,0,20,0,0,0,0,0,0,0,12,0,22,0,6,0,5,0,8,0,12,0,12,0,0,0,0,3,4,0,24,0,0,0,112,0,0,0,0,0,0,0,0,0,10,0,24,0,12,0,4,0,8,0,10,0,0,0,220,0,0,0,16,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,12,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,16,0,0,0,0,0,0,0,5,0,0,0,0,0,0,0,24,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,24,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,0,0,0,0,0,0,0,12,0,0,0,0,0,0,0,48,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,48,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,48,0,0,0,0,0,0,0,64,0,0,0,0,0,0,0,0,0,0,0,6,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,5,0,0,0,110,97,109,101,50,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,2,0,0,0,4,0,0,0,0,0,0,0,0,0,0,0,0,0,240,63,0,0,0,0,0,0,240,63,0,0,0,0,0,0,0,64,0,0,0,0,0,0,0,64,0,0,0,0,0,0,8,64,0,0,0,0,0,0,8,64,0,0,0,0,0,0,16,64,0,0,0,0,0,0,16,64,255,255,255,255,0,0,0,0,20,0,0,0,0,0,14,0,24,0,6,0,8,0,12,0,16,0,20,0,14,0,0,0,0,0,4,0,132,2,0,0,108,2,0,0,72,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,208,250,255,255,36,2,0,0,4,0,0,0,20,2,0,0,123,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,109,117,108,116,105,108,105,110,101,115,116,114,105,110,103,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,98,98,111,120,34,58,91,49,46,48,44,49,46,48,44,52,46,48,44,52,46,48,93,125,125,125,0,0,0,0,8,0,0,0,103,100,97,108,58,103,101,111,0,0,0,0,1,0,0,0,96,4,0,0,0,0,0,0,128,1,0,0,0,0,0,0,112,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,10,0,12,0,0,0,4,0,8,0,10,0,0,0,100,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,80,253,255,255,72,2,0,0,4,0,0,0,59,2,0,0,123,34,115,99,104,101,109,97,95,118,101,114,115,105,111,110,34,58,34,48,46,49,46,48,34,44,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,109,117,108,116,105,108,105,110,101,115,116,114,105,110,103,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,103,100,97,108,58,103,101,111,109,101,116,114,121,95,116,121,112,101,34,58,34,77,117,108,116,105,76,105,110,101,83,116,114,105,110,103,34,125,125,125,0,3,0,0,0,103,101,111,0,3,0,0,0,136,1,0,0,72,1,0,0,24,0,0,0,0,0,18,0,24,0,8,0,6,0,7,0,12,0,0,0,16,0,20,0,18,0,0,0,0,0,1,12,24,0,0,0,124,0,0,0,8,0,0,0,28,0,0,0,1,0,0,0,112,0,0,0,8,0,0,0,103,101,111,109,101,116,114,121,0,0,0,0,1,0,0,0,12,0,0,0,8,0,12,0,4,0,8,0,8,0,0,0,40,0,0,0,4,0,0,0,24,0,0,0,103,101,111,97,114,114,111,119,46,109,117,108,116,105,108,105,110,101,115,116,114,105,110,103,0,0,0,0,20,0,0,0,65,82,82,79,87,58,101,120,116,101,110,115,105,111,110,58,110,97,109,101,0,0,0,0,56,255,255,255,52,255,255,255,0,0,1,12,20,0,0,0,28,0,0,0,4,0,0,0,1,0,0,0,20,0,0,0,4,0,0,0,105,116,101,109,0,0,0,0,100,255,255,255,96,255,255,255,0,0,1,16,20,0,0,0,32,0,0,0,4,0,0,0,1,0,0,0,44,0,0,0,4,0,0,0,105,116,101,109,0,0,6,0,8,0,4,0,6,0,0,0,2,0,0,0,16,0,20,0,8,0,0,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,0,3,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,120,121,0,0,0,0,6,0,8,0,6,0,6,0,0,0,0,0,2,0,212,255,255,255,0,0,1,5,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,4,0,0,0,110,97,109,101,0,0,0,0,4,0,4,0,4,0,0,0,16,0,20,0,8,0,6,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,1,2,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,105,100,0,0,8,0,12,0,8,0,7,0,8,0,0,0,0,0,0,1,32,0,0,0,200,6,0,0,65,82,82,79,87,49';
 const expectedMultiLineStringGeoJson = {
@@ -149,7 +149,7 @@ const expectedMultiLineStringGeoJson = {
   ]
 };
 
-// a simple geojson file contains one MultiPolygon
+// a simple geojson contains one MultiPolygon
 const testMultiPolygonArrowRawBytes =
   '65,82,82,79,87,49,0,0,255,255,255,255,120,4,0,0,16,0,0,0,0,0,10,0,14,0,6,0,5,0,8,0,10,0,0,0,0,1,4,0,16,0,0,0,0,0,10,0,12,0,0,0,4,0,8,0,10,0,0,0,96,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,84,253,255,255,68,2,0,0,4,0,0,0,53,2,0,0,123,34,115,99,104,101,109,97,95,118,101,114,115,105,111,110,34,58,34,48,46,49,46,48,34,44,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,109,117,108,116,105,112,111,108,121,103,111,110,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,103,100,97,108,58,103,101,111,109,101,116,114,121,95,116,121,112,101,34,58,34,77,117,108,116,105,80,111,108,121,103,111,110,34,125,125,125,0,0,0,3,0,0,0,103,101,111,0,3,0,0,0,176,1,0,0,112,1,0,0,24,0,0,0,0,0,18,0,24,0,8,0,6,0,7,0,12,0,0,0,16,0,20,0,18,0,0,0,0,0,1,12,24,0,0,0,120,0,0,0,8,0,0,0,28,0,0,0,1,0,0,0,108,0,0,0,8,0,0,0,103,101,111,109,101,116,114,121,0,0,0,0,1,0,0,0,12,0,0,0,8,0,12,0,4,0,8,0,8,0,0,0,36,0,0,0,4,0,0,0,21,0,0,0,103,101,111,97,114,114,111,119,46,109,117,108,116,105,112,111,108,121,103,111,110,0,0,0,20,0,0,0,65,82,82,79,87,58,101,120,116,101,110,115,105,111,110,58,110,97,109,101,0,0,0,0,12,255,255,255,8,255,255,255,0,0,1,12,20,0,0,0,28,0,0,0,4,0,0,0,1,0,0,0,20,0,0,0,4,0,0,0,105,116,101,109,0,0,0,0,56,255,255,255,52,255,255,255,0,0,1,12,20,0,0,0,28,0,0,0,4,0,0,0,1,0,0,0,20,0,0,0,4,0,0,0,105,116,101,109,0,0,0,0,100,255,255,255,96,255,255,255,0,0,1,16,20,0,0,0,32,0,0,0,4,0,0,0,1,0,0,0,44,0,0,0,4,0,0,0,105,116,101,109,0,0,6,0,8,0,4,0,6,0,0,0,2,0,0,0,16,0,20,0,8,0,0,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,0,3,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,120,121,0,0,0,0,6,0,8,0,6,0,6,0,0,0,0,0,2,0,212,255,255,255,0,0,1,5,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,4,0,0,0,110,97,109,101,0,0,0,0,4,0,4,0,4,0,0,0,16,0,20,0,8,0,6,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,1,2,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,105,100,0,0,8,0,12,0,8,0,7,0,8,0,0,0,0,0,0,1,32,0,0,0,0,0,0,0,255,255,255,255,168,1,0,0,20,0,0,0,0,0,0,0,12,0,22,0,6,0,5,0,8,0,12,0,12,0,0,0,0,3,4,0,24,0,0,0,224,0,0,0,0,0,0,0,0,0,10,0,24,0,12,0,4,0,8,0,10,0,0,0,252,0,0,0,16,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,14,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,16,0,0,0,0,0,0,0,5,0,0,0,0,0,0,0,24,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,24,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,0,0,0,0,0,0,0,12,0,0,0,0,0,0,0,48,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,48,0,0,0,0,0,0,0,12,0,0,0,0,0,0,0,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,64,0,0,0,0,0,0,0,160,0,0,0,0,0,0,0,0,0,0,0,7,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,10,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,20,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,5,0,0,0,110,97,109,101,49,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,1,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,5,0,0,0,10,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,240,63,0,0,0,0,0,0,240,63,0,0,0,0,0,0,240,63,0,0,0,0,0,0,240,63,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,64,0,0,0,0,0,0,0,64,0,0,0,0,0,0,0,64,0,0,0,0,0,0,8,64,0,0,0,0,0,0,8,64,0,0,0,0,0,0,8,64,0,0,0,0,0,0,8,64,0,0,0,0,0,0,0,64,0,0,0,0,0,0,0,64,0,0,0,0,0,0,0,64,255,255,255,255,0,0,0,0,24,0,0,0,0,0,0,0,0,0,14,0,24,0,6,0,8,0,12,0,16,0,20,0,14,0,0,0,0,0,4,0,124,2,0,0,104,2,0,0,68,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,220,250,255,255,32,2,0,0,4,0,0,0,17,2,0,0,123,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,109,117,108,116,105,112,111,108,121,103,111,110,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,98,98,111,120,34,58,91,48,46,48,44,48,46,48,44,51,46,48,44,51,46,48,93,125,125,125,0,0,0,8,0,0,0,103,100,97,108,58,103,101,111,0,0,0,0,1,0,0,0,136,4,0,0,0,0,0,0,176,1,0,0,0,0,0,0,224,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,10,0,12,0,0,0,4,0,8,0,10,0,0,0,96,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,84,253,255,255,68,2,0,0,4,0,0,0,53,2,0,0,123,34,115,99,104,101,109,97,95,118,101,114,115,105,111,110,34,58,34,48,46,49,46,48,34,44,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,103,101,111,97,114,114,111,119,46,109,117,108,116,105,112,111,108,121,103,111,110,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,103,100,97,108,58,103,101,111,109,101,116,114,121,95,116,121,112,101,34,58,34,77,117,108,116,105,80,111,108,121,103,111,110,34,125,125,125,0,0,0,3,0,0,0,103,101,111,0,3,0,0,0,176,1,0,0,112,1,0,0,24,0,0,0,0,0,18,0,24,0,8,0,6,0,7,0,12,0,0,0,16,0,20,0,18,0,0,0,0,0,1,12,24,0,0,0,120,0,0,0,8,0,0,0,28,0,0,0,1,0,0,0,108,0,0,0,8,0,0,0,103,101,111,109,101,116,114,121,0,0,0,0,1,0,0,0,12,0,0,0,8,0,12,0,4,0,8,0,8,0,0,0,36,0,0,0,4,0,0,0,21,0,0,0,103,101,111,97,114,114,111,119,46,109,117,108,116,105,112,111,108,121,103,111,110,0,0,0,20,0,0,0,65,82,82,79,87,58,101,120,116,101,110,115,105,111,110,58,110,97,109,101,0,0,0,0,12,255,255,255,8,255,255,255,0,0,1,12,20,0,0,0,28,0,0,0,4,0,0,0,1,0,0,0,20,0,0,0,4,0,0,0,105,116,101,109,0,0,0,0,56,255,255,255,52,255,255,255,0,0,1,12,20,0,0,0,28,0,0,0,4,0,0,0,1,0,0,0,20,0,0,0,4,0,0,0,105,116,101,109,0,0,0,0,100,255,255,255,96,255,255,255,0,0,1,16,20,0,0,0,32,0,0,0,4,0,0,0,1,0,0,0,44,0,0,0,4,0,0,0,105,116,101,109,0,0,6,0,8,0,4,0,6,0,0,0,2,0,0,0,16,0,20,0,8,0,0,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,0,3,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,120,121,0,0,0,0,6,0,8,0,6,0,6,0,0,0,0,0,2,0,212,255,255,255,0,0,1,5,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,4,0,0,0,110,97,109,101,0,0,0,0,4,0,4,0,4,0,0,0,16,0,20,0,8,0,6,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,1,2,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,105,100,0,0,8,0,12,0,8,0,7,0,8,0,0,0,0,0,0,1,32,0,0,0,232,6,0,0,65,82,82,79,87,49';
 const expectedMultiPolygonGeojson = {
@@ -196,13 +196,15 @@ const testPointArrowWKTRawBytes =
 const testPointArrowWKBRawBytes =
   '65,82,82,79,87,49,0,0,255,255,255,255,120,3,0,0,16,0,0,0,0,0,10,0,14,0,6,0,5,0,8,0,10,0,0,0,0,1,4,0,16,0,0,0,0,0,10,0,12,0,0,0,4,0,8,0,10,0,0,0,72,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,112,253,255,255,44,2,0,0,4,0,0,0,28,2,0,0,123,34,115,99,104,101,109,97,95,118,101,114,115,105,111,110,34,58,34,48,46,49,46,48,34,44,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,87,75,66,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,103,100,97,108,58,103,101,111,109,101,116,114,121,95,116,121,112,101,34,58,34,80,111,105,110,116,34,125,125,125,0,0,0,0,3,0,0,0,103,101,111,0,3,0,0,0,204,0,0,0,140,0,0,0,24,0,0,0,0,0,18,0,24,0,8,0,6,0,7,0,12,0,0,0,16,0,20,0,18,0,0,0,0,0,1,4,20,0,0,0,96,0,0,0,8,0,0,0,24,0,0,0,0,0,0,0,8,0,0,0,103,101,111,109,101,116,114,121,0,0,0,0,1,0,0,0,12,0,0,0,8,0,12,0,4,0,8,0,8,0,0,0,16,0,0,0,4,0,0,0,3,0,0,0,87,75,66,0,20,0,0,0,65,82,82,79,87,58,101,120,116,101,110,115,105,111,110,58,110,97,109,101,0,0,0,0,216,255,255,255,212,255,255,255,0,0,1,5,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,4,0,0,0,110,97,109,101,0,0,0,0,4,0,4,0,4,0,0,0,16,0,20,0,8,0,6,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,1,2,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,105,100,0,0,8,0,12,0,8,0,7,0,8,0,0,0,0,0,0,1,32,0,0,0,255,255,255,255,8,1,0,0,20,0,0,0,0,0,0,0,12,0,22,0,6,0,5,0,8,0,12,0,12,0,0,0,0,3,4,0,24,0,0,0,56,0,0,0,0,0,0,0,0,0,10,0,24,0,12,0,4,0,8,0,10,0,0,0,156,0,0,0,16,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,16,0,0,0,0,0,0,0,5,0,0,0,0,0,0,0,24,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,24,0,0,0,0,0,0,0,8,0,0,0,0,0,0,0,32,0,0,0,0,0,0,0,21,0,0,0,0,0,0,0,0,0,0,0,3,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,5,0,0,0,110,97,109,101,49,0,0,0,0,0,0,0,21,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,255,255,255,255,0,0,0,0,20,0,0,0,0,0,14,0,24,0,6,0,8,0,12,0,16,0,20,0,14,0,0,0,0,0,4,0,108,2,0,0,84,2,0,0,48,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,8,251,255,255,12,2,0,0,4,0,0,0,255,1,0,0,123,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,87,75,66,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,98,98,111,120,34,58,91,48,46,48,44,48,46,48,44,48,46,48,44,48,46,48,93,125,125,125,0,8,0,0,0,103,100,97,108,58,103,101,111,0,0,0,0,1,0,0,0,136,3,0,0,0,0,0,0,16,1,0,0,0,0,0,0,56,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,10,0,12,0,0,0,4,0,8,0,10,0,0,0,72,2,0,0,4,0,0,0,1,0,0,0,4,0,0,0,112,253,255,255,44,2,0,0,4,0,0,0,28,2,0,0,123,34,115,99,104,101,109,97,95,118,101,114,115,105,111,110,34,58,34,48,46,49,46,48,34,44,34,112,114,105,109,97,114,121,95,99,111,108,117,109,110,34,58,34,103,101,111,109,101,116,114,121,34,44,34,99,111,108,117,109,110,115,34,58,123,34,103,101,111,109,101,116,114,121,34,58,123,34,101,110,99,111,100,105,110,103,34,58,34,87,75,66,34,44,34,99,114,115,34,58,34,71,69,79,71,67,82,83,91,92,34,87,71,83,32,56,52,92,34,44,68,65,84,85,77,91,92,34,87,111,114,108,100,32,71,101,111,100,101,116,105,99,32,83,121,115,116,101,109,32,49,57,56,52,92,34,44,69,76,76,73,80,83,79,73,68,91,92,34,87,71,83,32,56,52,92,34,44,54,51,55,56,49,51,55,44,50,57,56,46,50,53,55,50,50,51,53,54,51,44,76,69,78,71,84,72,85,78,73,84,91,92,34,109,101,116,114,101,92,34,44,49,93,93,93,44,80,82,73,77,69,77,91,92,34,71,114,101,101,110,119,105,99,104,92,34,44,48,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,67,83,91,101,108,108,105,112,115,111,105,100,97,108,44,50,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,97,116,105,116,117,100,101,32,40,76,97,116,41,92,34,44,110,111,114,116,104,44,79,82,68,69,82,91,49,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,65,88,73,83,91,92,34,103,101,111,100,101,116,105,99,32,108,111,110,103,105,116,117,100,101,32,40,76,111,110,41,92,34,44,101,97,115,116,44,79,82,68,69,82,91,50,93,44,65,78,71,76,69,85,78,73,84,91,92,34,100,101,103,114,101,101,92,34,44,48,46,48,49,55,52,53,51,50,57,50,53,49,57,57,52,51,51,93,93,44,73,68,91,92,34,69,80,83,71,92,34,44,52,51,50,54,93,93,34,44,34,103,100,97,108,58,103,101,111,109,101,116,114,121,95,116,121,112,101,34,58,34,80,111,105,110,116,34,125,125,125,0,0,0,0,3,0,0,0,103,101,111,0,3,0,0,0,204,0,0,0,140,0,0,0,24,0,0,0,0,0,18,0,24,0,8,0,6,0,7,0,12,0,0,0,16,0,20,0,18,0,0,0,0,0,1,4,20,0,0,0,96,0,0,0,8,0,0,0,24,0,0,0,0,0,0,0,8,0,0,0,103,101,111,109,101,116,114,121,0,0,0,0,1,0,0,0,12,0,0,0,8,0,12,0,4,0,8,0,8,0,0,0,16,0,0,0,4,0,0,0,3,0,0,0,87,75,66,0,20,0,0,0,65,82,82,79,87,58,101,120,116,101,110,115,105,111,110,58,110,97,109,101,0,0,0,0,216,255,255,255,212,255,255,255,0,0,1,5,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,4,0,0,0,110,97,109,101,0,0,0,0,4,0,4,0,4,0,0,0,16,0,20,0,8,0,6,0,7,0,12,0,0,0,16,0,16,0,0,0,0,0,1,2,16,0,0,0,28,0,0,0,4,0,0,0,0,0,0,0,2,0,0,0,105,100,0,0,8,0,12,0,8,0,7,0,8,0,0,0,0,0,0,1,32,0,0,0,216,5,0,0,65,82,82,79,87,49';
 
-  // a simple geojson file contains one point
+  // a simple geojson contains one point
 const expectedPointGeojsonForWellKnownFormat = {
   type: 'FeatureCollection',
   features: [
     {
       type: 'Feature',
       properties: {
+        id: 1,
+        name: 'name1'
       },
       geometry: {
         type: 'Point',
@@ -250,7 +252,10 @@ test.only('geojsonUtils.parseGeometryFromArrow', async t => {
     const encoding = geometryColumns['geometry'].encoding;
 
     // check encoding is one of GEOARROW_ENCODINGS
-    t.ok(encoding.startsWith('geoarrow'), 'encoding is one of GEOARROW_ENCODINGS');
+    t.ok(
+      Object.values(GEOJSON_ENCODINGS).includes(encoding),
+      'encoding is one of GEOARROW_ENCODINGS'
+    );
 
     // get first geometry from arrow geometry column
     const firstArrowGeometry = arrowTable.getColumn('geometry').get(0);
